Rename core/messages to core/message
After talking to xla, singular form is the norm in Rust and it also
helps unifying the core and the client message modules.
