+++ client/examples/transaction_signing.rs
@@ -3,7 +3,7 @@ use futures::compat::Compat;
 use futures::future::FutureExt;
 
 use radicle_registry_client::{
-    ed25519, messages::Transfer, Client, ClientT as _, CryptoPair as _, Error, Transaction,
+    ed25519, message::Transfer, Client, ClientT as _, CryptoPair as _, Error, Transaction,
     TransactionExtra,
 };
