+++ core/src/message.rs
